% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationfunnel-package.R
\docType{package}
\name{correlationfunnel-package}
\alias{correlationfunnel}
\alias{correlationfunnel-package}
\title{correlationfunnel: Speed Up Exploratory Data Analysis (EDA) with the Correlation Funnel}
\description{
The 'correlationfunnel' package drastically speeds up Exploratory Data Analysis (EDA)
    by providing a succinct workflow and interactive visualization tools for understanding
    which features have relationships to target (response). This is excellent for pre-modeling / 
    pre-Machine Learning since you can determine quickly if you have predictive features, 
    those with relationship to the target feature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/business-science/correlationfunnel}
  \item Report bugs at \url{https://github.com/business-science/correlationfunnel/issues}
}

}
\author{
\strong{Maintainer}: Matt Dancho \email{mdancho@business-science.io}

}
\keyword{internal}
