% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer_churn_tbl}
\alias{customer_churn_tbl}
\title{Customer Churn Data Set for a Telecommunications Company}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7043 rows and 21 columns.}
\source{
\href{https://community.ibm.com/community/user/gettingstarted/home}{IBM Sample Datasets}
}
\usage{
customer_churn_tbl
}
\description{
A dataset containing data related to telecom customers that have enrolled in various products and services
}
\details{
Telecom Customer Data:
\itemize{
\item customerID (chr): CUSTOMER ID
\item gender (chr): Customer's gender ("Female", "Male")
\item SeniorCitizen (dbl): 1 = Senior Citzen, 0 = Not Senior Citizen
\item Partner (chr): Whether the customer has a partner or not (Yes, No)
\item Dependents (chr): Whether the customer has dependents or not (Yes, No)
\item tenure (dbl): Number of months the customer has stayed with the company
\item PhoneService (chr): Whether the customer has a phone service or not (Yes, No)
\item MultipleLines (chr): Whether the customer has multiple lines or not (Yes, No, No phone service)
\item InternetService (chr): Customer’s internet service provider (DSL, Fiber optic, No)
\item OnlineSecurity (chr): Whether the customer has online security or not (Yes, No, No internet service)
\item OnlineBackup (chr): Whether the customer has online backup or not (Yes, No, No internet service)
\item DeviceProtection (chr): Whether the customer has device protection or not (Yes, No, No internet service)
\item TechSupport (chr): Whether the customer has tech support or not (Yes, No, No internet service)
\item StreamingTV (chr): Whether the customer has streaming TV or not (Yes, No, No internet service)
\item StreamingMovies (chr): Whether the customer has streaming movies or not (Yes, No, No internet service)
\item Contract (chr): The contract term of the customer (Month-to-month, One year, Two year)
\item PaperlessBilling (chr): Whether the customer has paperless billing or not (Yes, No)
\item PaymentMethod (chr): The customer’s payment method (Electronic check, Mailed check, Bank transfer (automatic), Credit card (automatic))
\item MonthlyCharges (dbl): The amount charged to the customer monthly
\item TotalCharges (dbl): The total amount charged to the customer
\item Churn (chr): Outcome. Whether the customer churned or not (Yes or No)
}
}
\keyword{datasets}
