% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationfunnel-package.R
\docType{package}
\name{correlationfunnel-package}
\alias{correlationfunnel}
\alias{correlationfunnel-package}
\title{correlationfunnel: Speed Up Exploratory Data Analysis (EDA) with the Correlation Funnel}
\description{
Speeds up exploratory data analysis (EDA)
    by providing a succinct workflow and interactive visualization tools for understanding
    which features have relationships to target (response). Uses binary correlation analysis
    to determine relationship. Default correlation method is the Pearson method. 
    Lian Duan, W Nick Street, Yanchi Liu, Songhua Xu, and Brook Wu (2014) <doi:10.1145/2637484>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/business-science/correlationfunnel}
  \item Report bugs at \url{https://github.com/business-science/correlationfunnel/issues}
}

}
\author{
\strong{Maintainer}: Matt Dancho \email{mdancho@business-science.io}

}
\keyword{internal}
